+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+<a name="4.3.1"></a>
+# [4.3.1](https://github.com/glowyphp/strings) (2022-05-08)
+* Fix toNull() method.
+
+<a name="4.3.0"></a>
+# [4.3.0](https://github.com/glowyphp/strings) (2022-05-08)
+* Added isNull() method.
+* Added isInteger() method.
+* Added isFloat() method.
+* Added isUuid() method.
+* Added toNull() method.
+* Method pipe() returns a new instance of the class.
+
 <a name="4.2.0"></a>
 # [4.2.0](https://github.com/glowyphp/strings) (2022-01-01)
 * Added isHexColor() method.